Implement iterateValidDays generator and refactor getNextExecution for improved date handling
