+++ backend/src/scheduler/calculator/next.js
@@ -3,6 +3,7 @@
  */
 
 const { dateTimeFromObject } = require('../../datetime');
+const { iterateValidDays } = require('../expression');
 const { matchesCronExpression } = require('./current');
 
 /**
@@ -13,78 +14,62 @@ const { matchesCronExpression } = require('./current');
  * @throws {CronCalculationError} If next execution cannot be calculated
  */
 function getNextExecution(cronExpr, origin) {
-    let year = origin.year;
-    let month = origin.month;
+    for (const { year, month, day } of iterateValidDays(cronExpr, origin)) {
+        const getTime = () => {
+            if (day === origin.day && year === origin.year && month === origin.month) {
+                const hour = cronExpr.validHours.filter(h => h >= origin.hour)[0];
+                if (hour === undefined) {
+                    return null;
+                }
 
-    // eslint-disable-next-line no-constant-condition
-    while (true) {
-        let validDays = cronExpr.validDays(year, month);
-        if (year === origin.year && month === origin.month) {
-            validDays = validDays.filter(d => d >= origin.day);
-        }
-
-        for (const day of validDays) {
-            const getTime = () => {
-                if (day === origin.day && year === origin.year && month === origin.month) {
-                    const hour = cronExpr.validHours.filter(h => h >= origin.hour)[0];
+                const minute = hour === origin.hour
+                    ? cronExpr.validMinutes.filter(m => m > origin.minute)[0]
+                    : cronExpr.validMinutes[0];
+                if (minute === undefined) {
+                    const hour = cronExpr.validHours.filter(h => h > origin.hour)[0];
                     if (hour === undefined) {
                         return null;
                     }
-
-                    const minute = hour === origin.hour
-                        ? cronExpr.validMinutes.filter(m => m > origin.minute)[0]
-                        : cronExpr.validMinutes[0];
+                    const minute = cronExpr.validMinutes[0];
                     if (minute === undefined) {
-                        const hour = cronExpr.validHours.filter(h => h > origin.hour)[0];
-                        if (hour === undefined) {
-                            return null;
-                        }
-                        const minute = cronExpr.validMinutes[0];
-                        if (minute === undefined) {
-                            return null;
-                        }
-                        return { hour, minute };
+                        return null;
                     }
-
                     return { hour, minute };
-                } else {
-                    return { hour: cronExpr.validHours[0], minute: cronExpr.validMinutes[0] };
                 }
-            };
 
-            const time = getTime();
-            if (time === null) {
-                continue;
-            }
-
-            const { hour, minute } = time;
-
-            const candidate = dateTimeFromObject({
-                year,
-                month,
-                day,
-                hour,
-                minute,
-                second: 0,
-                millisecond: 0,
-            });
-            if (candidate.isValid === false) {
-                throw new Error(`Invalid candidate datetime: ${candidate}`);
-            }
-            if (matchesCronExpression(cronExpr, candidate)) {
-                return candidate;
+                return { hour, minute };
             } else {
-                throw new Error("Internal error: candidate does not match cron expression");
+                return { hour: cronExpr.validHours[0], minute: cronExpr.validMinutes[0] };
             }
+        };
+
+        const time = getTime();
+        if (time === null) {
+            continue;
         }
 
-        // Advance to next month.
-        month += 1;
-        if (month > 12) {
-            month = 1;
-            year += 1;
+        const { hour, minute } = time;
+
+        const candidate = dateTimeFromObject({
+            year,
+            month,
+            day,
+            hour,
+            minute,
+            second: 0,
+            millisecond: 0,
+        });
+        if (candidate.isValid === false) {
+            throw new Error(`Invalid candidate datetime: ${candidate}`);
+        }
+        if (matchesCronExpression(cronExpr, candidate)) {
+            return candidate;
+        } else {
+            throw new Error("Internal error: candidate does not match cron expression");
         }
     }
+
+    throw new Error("No valid next execution time found");
 }
 
 module.exports = {
