+++ backend/src/scheduler/expression/structure.js
@@ -271,8 +271,44 @@ function parseCronExpression(expression) {
     }
 }
 
+
+/**
+ * Generator that yields valid (year, month, day) tuples starting from the given date.
+ * @param {CronExpression} cronExpr
+ * @param {import("../../datetime").DateTime} startDate
+ * @returns {Generator<{year: number, month: number, day: number}>} Tuples
+ */
+function* iterateValidDays(cronExpr, startDate) {
+    const origin = startDate;
+
+    const oyear = origin.year;
+    const omonth = origin.month;
+    let year = origin.year;
+    let month = origin.month;
+
+    // eslint-disable-next-line no-constant-condition
+    while (true) {
+        let validDays = cronExpr.validDays(year, month);
+        if (month === omonth && year === oyear) {
+            validDays = validDays.filter(d => d >= origin.day);
+        }
+
+        for (const day of validDays) {
+            yield { year, month, day };
+        }
+
+        month += 1;
+        if (month > 12) {
+            month = 1;
+            year += 1;
+        }
+    }
+}
+
+
 module.exports = {
     parseCronExpression,
     isCronExpression,
     isInvalidCronExpressionError,
+    iterateValidDays,
 };
